+++ GramsG4/gramsg4.cc
@@ -44,6 +44,11 @@
 #include "G4RunManager.hh"
 #endif
 
+
+// Solely for setting the verbosity of the physics display.
+#include "G4EmParameters.hh"
+#include "G4HadronicParameters.hh"
+
 #include "G4UImanager.hh"
 #include "G4VisExecutive.hh"
 #include "G4UIExecutive.hh"
@@ -201,6 +206,17 @@ int main(int argc,char **argv)
       exit(EXIT_FAILURE);
     }
 
+  // Setting whether the details of the physics list are displayed
+  // at the start of the job.
+  auto emParam = G4EmParameters::Instance();
+  auto hadParam = G4HadronicParameters::Instance();
+  emParam->SetVerbose(0);
+  hadParam->SetVerboseLevel(0);
+  if (verbose) {
+    emParam->SetVerbose(1);
+    hadParam->SetVerboseLevel(1);
+  }
+  
   // Include optical photons if requested.
   // This was copied from 
   // artg4tk/artg4tk/pluginActions/physicsList/PhysicsList_service.cc 
@@ -254,6 +270,11 @@ int main(int argc,char **argv)
     
   } // if opticalphysics
 
+  // Control the verbosity of the physics-list display at the
+  // start of the run.
+  
+
+  
   // Usual initializations: detector, physics, and user actions.
   runManager->SetUserInitialization(new gramsg4::DetectorConstruction());
   runManager->SetUserInitialization(physics);
@@ -279,7 +300,7 @@ int main(int argc,char **argv)
   // ***** end User Action Manager setup *****
 
   // Tell Geant4 we're about to begin. 
-  runManager->Initialize();
+  // runManager->Initialize();
 
   // Initialize visualization
   G4VisManager* visManager = new G4VisExecutive;